delete overlap table only if upserting (created)

+++ src/main/java/bio/terra/flight/dataset/ingest/IngestCleanupStep.java
@@ -1,6 +1,7 @@
 package bio.terra.flight.dataset.ingest;
 
 import bio.terra.metadata.Dataset;
+import bio.terra.model.IngestRequestModel;
 import bio.terra.pdao.bigquery.BigQueryPdao;
 import bio.terra.service.DatasetService;
 import bio.terra.stairway.FlightContext;
@@ -37,10 +38,15 @@ public class IngestCleanupStep implements Step {
         }
 
         try {
-            Dataset dataset = IngestUtils.getDataset(context, datasetService);
+            IngestRequestModel ingestRequestModel = IngestUtils.getIngestRequestModel(context);
+            IngestRequestModel.StrategyEnum ingestStrategy = ingestRequestModel.getStrategy();
+            if (ingestStrategy == IngestRequestModel.StrategyEnum.UPSERT) {
+                Dataset dataset = IngestUtils.getDataset(context, datasetService);
 
-            overlappingTableName = IngestUtils.getOverlappingTableName(context);
-            bigQueryPdao.deleteDatasetTable(dataset, overlappingTableName);
+                overlappingTableName = IngestUtils.getOverlappingTableName(context);
+
+                bigQueryPdao.deleteDatasetTable(dataset, overlappingTableName);
+            }
         } catch (Exception ex) {
             logger.error("Failure deleting overlapping table: " + overlappingTableName, ex);
         }
